refactor: simplify local Stagehand code-mode tool
