+++ packages/stagehand-codemode/package.json
@@ -15,12 +15,9 @@
   },
   "scripts": {
     "build": "tsdown",
-    "test": "vitest run",
-    "test:live": "pnpm run build && node tests/live-mcp-smoke.mjs",
     "typecheck": "tsc --noEmit"
   },
   "dependencies": {
-    "@browserbasehq/sdk": "^2.16.0",
     "@modelcontextprotocol/sdk": "^1.30.0",
     "zod": "^4.4.3"
   },
@@ -35,8 +32,7 @@
   "devDependencies": {
     "@types/node": "^25.0.9",
     "tsdown": "^0.15.4",
-    "typescript": "^5.9.3",
-    "vitest": "^4.0.6"
+    "typescript": "^5.9.3"
   },
   "engines": {
     "node": ">=22.18.0"
