+++ packages/stagehand-codemode/src/tool-contract.ts
@@ -5,7 +5,8 @@ import type { CodeExecuteResult } from './types.js';
 export const CODE_EXECUTE_DESCRIPTION = [
   'Execute an async JavaScript function body against one long-lived Stagehand V4 browser on Browserbase.',
   'The local executor lazily creates the browser on the first call and reuses it for later calls.',
-  'This runs trusted local code with local filesystem, network, and in-process SDK access; it is not a security sandbox.',
+  'Code runs directly in the local MCP process. The owning agent framework should terminate and restart that process if it stops responding.',
+  'This is trusted local code with filesystem, network, and in-process SDK access; it is not a security sandbox.',
   '',
   STAGEHAND_CODEMODE_SKILL,
 ].join('\n');
@@ -18,13 +19,6 @@ export const codeExecuteSchema = z.object({
     .describe(
       'Async JavaScript function body. page, context, stagehand, z, and console are in scope.'
     ),
-  timeout_ms: z
-    .number()
-    .int()
-    .positive()
-    .max(300_000)
-    .optional()
-    .describe('Per-call execution timeout in milliseconds.'),
 });
 
 export function codeExecuteResultText(result: CodeExecuteResult): string {
