+++ packages/stagehand-codemode/tsdown.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from 'tsdown';
 
 export default defineConfig({
-  entry: ['src/index.ts', 'src/cli.ts', 'src/runtime-child.ts'],
+  entry: ['src/index.ts', 'src/cli.ts'],
   format: ['esm'],
   platform: 'node',
   target: 'node22',
